Extensive renaming
All variable and function names across the program should now use CO terminology to reflect the new functionality of the program (e.g. "Holodeck" and "Tribble" become "Live" and "Playtest respectively, "ValidateStoInstallationPath" becomes "ValidateCoInstallationPath" and so on). Various UI elements (both internally and externally) have also be renamed and edited to reflect the new functionality of the progam.